pyproject.toml
[tool.poetry]
name = "gapp"
version = "0.1.0"
description = ""
authors = ["ArthurGolubev <gg.arthur.go@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = {extras = ["standard"], version = "^0.18.2"}
neo4j = "^4.4.4"
loguru = "^0.6.0"
py2neo = "^2021.2.3"
strawberry-graphql = {extras = ["fastapi"], version = "^0.117.1"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
